requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
starlette==0.37.2
slowapi==0.1.9
python-dotenv>=1.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
joblib>=1.3.0
# PostgreSQL support (uncomment when deploying with PostgreSQL)
# psycopg2-binary>=2.9.0
